Prevents update methods from creating new entities.
The Identifiable interface has been introduced and implemented within various entities to better handle object identification. Logic has been added to the update method in the AbstractDataService to throw an UnsupportedOperationException in case attempts are made to change the ID. A handler for this exception has also been added to the GlobalExceptionHandlerController. This prevents accidental creation of new elements when intending to update existing ones.